backend/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-filter==24.2
psycopg2
django-jazzmin
drf-spectacular==0.27.2
celery~=5.4.0
redis
Pillow
